Developments/requirements.txt
transformers
fastapi
uvicorn
python-dotenv
pydantic
requests
accelerate